// src/app/common/bpmn/palette/palette.enum.ts
export enum PaletteCategory {
  annotation = 'annotation', // or note
  boundary = 'boundary',
  event = 'event', // start, intermedite and end
  activity = 'activity', // or tasks
  dataObject = 'dataobject',
  dataWarehousing = 'datastore',
  gateway = 'gateway',
  group = 'group',
  pool = 'Pool',
  subprocess = 'subprocess',
  lane = 'Lane',
  header = 'header',
  privateProcess = 'privateProcess',
  msg = 'msg',
  data = 'data',
  diagram = 'diagram',
  link = 'Link'
}